requirements.txt
Flask==0.12.2
gunicorn==19.7.1
pymongo==3.5.1
Pillow==4.1.0
PyYAML==3.12
Flask-RESTful==0.3.6
Flask-JWT==0.3.2
requests==2.18.4